planner: move rule_elimate_projection into rules package
