+++ pkg/planner/core/rule/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "rule_collect_plan_stats.go",
         "rule_column_pruning.go",
         "rule_constant_propagation.go",
+        "rule_eliminate_projection.go",
         "rule_init.go",
         "rule_max_min_eliminate.go",
         "rule_outer_join_to_semi_join.go",
